+++ src/lib/colors.js
@@ -1,0 +1,9 @@
+const colors = {
+  black: '#131415',
+  white: '#fff',
+  pink: '#ff1493',
+  gray: '#dedede',
+  red: '#E74C3C',
+}
+
+export default colors
